requirements.txt
alembic==1.4.2
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.6
Flask-Ext==0.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.0
WTForms==2.2.1
